agents/python/execute_with_memories/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.7.2
python-dotenv==1.0.1
requests==2.31.0

# MCP dependencies
mcp==1.6.0
httpx==0.27.0
anyio>=3.0 # Often a dependency of httpx/fastapi
trio>=0.20 # Add trio explicitly

# Optional dependencies
colorama==0.4.6